Exception handlers

+++ src/main/java/com/hakkinenT/dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.hakkinenT.dscommerce.services;
 import com.hakkinenT.dscommerce.dto.ProductDTO;
 import com.hakkinenT.dscommerce.entities.Product;
 import com.hakkinenT.dscommerce.repositories.ProductRepository;
+import com.hakkinenT.dscommerce.services.exceptions.DatabaseException;
+import com.hakkinenT.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -19,8 +24,9 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Optional<Product> result = repository.findById(id);
-        Product product = result.get();
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
+
         ProductDTO dto = new ProductDTO(product);
         return dto;
     }
@@ -43,12 +49,16 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
+        try{
+            Product entity = repository.getReferenceById(id);
 
-        copyDtoToEntity(dto, entity);
+            copyDtoToEntity(dto, entity);
 
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
@@ -59,8 +69,16 @@ public class ProductService {
 
     }
 
-    @Transactional
-    public void delete(Long id){
-        repository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 }
